Add CMake option for forcing SpiderMonkey to be built from source
Before this, the build system would always check if prebuilt artifacts exist, unless pre-built artifacts were provided locally. With this, it's possible to force rebuilding the engine from source instead. That's useful for development as it ensures that debugging symbols with local paths are available, and makes it easy to make exploratory changes to the source.

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>