Use mozjs for SpiderMonkey integration
This moves us from StarlingMonkey-built SpiderMonkey bindings and build system stuff to mozjs. We might need some patches of our own on top of what upstream carries, but for now this is (a pending PR landing) unmodified upstream.

The CI-built archives for mozjs-sys unfortunately aren't usable right now, because the linker can't find `libwasi-emulated-getpid` when doing so. That shouldn't be too hard to fix, but in the meantime, defining `MOZJS_FROM_SOURCE=1` and being a bit more patience will address this.