Habemus event loop
This adds a whole lot of stuff, but it's all in service of making the event loop work.

Which it now does, in that timer events (i.e. `setTimeout` et al) work.

Right now there aren't really any other events that could work, since no IO APIs are hooked up, but that's for another day.