Add oh so much stuff
So this ... compiles? And actually runs as well as before. And has `{Readable,Writable,Transform}Stream` working. and `URL` et al. So that is all very good.

This does not, however, do anything actually related to networking as promised in the previous commit. Instead, it adds even more infrastructure for it (which kind of accidentally included the above builtins and [checks notes] 33 others.)